backend/requirements.txt
# FastAPI and web framework
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.27
asyncpg>=0.29.0
alembic>=1.13.1
psycopg[binary]>=3.2.0

# Celery and async tasks
celery>=5.3.6
redis>=5.0.1

# Data validation
pydantic>=2.6.1
pydantic-settings>=2.1.0

# Google Cloud
google-cloud-storage>=2.14.0

# Audio processing
yt-dlp>=2025.10.14
numpy<2.0.0  # Pin to 1.x for torch 2.2.0 compatibility
librosa>=0.10.1
soundfile>=0.12.1
torch==2.2.0  # Pin to version compatible with Demucs without torchcodec
torchaudio==2.2.0  # Must match torch version
demucs>=4.0.1
pydub>=0.25.1
mutagen>=1.47.0
av>=12.0.0  # PyAV for audio decoding

# Utilities
python-dotenv>=1.0.1
httpx>=0.26.0
aiofiles>=23.2.1

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

